Use explicit casts when dealing with global variables

+++ MWSE/TES3WorldController.cpp
@@ -382,10 +382,10 @@ namespace TES3 {
 		std::stringstream output;
 
 		// Build header.
-		int day = worldController->gvarDay->value;
-		const char* month = worldController->getNameForMonth(worldController->gvarMonth->value);
+		int day = static_cast<int>(worldController->gvarDay->value);
+		const char* month = worldController->getNameForMonth(static_cast<int>(worldController->gvarMonth->value));
 		const char* gmstDay = ndd->GMSTs[GMST::sDay]->value.asString;
-		int daysPassed = worldController->gvarDaysPassed->value;
+		int daysPassed = static_cast<int>(worldController->gvarDaysPassed->value);
 		output << "<FONT COLOR=\"9F0000\">";
 		output << day << " " << month;
 		output << " (" << gmstDay << " " << daysPassed << ")";
@@ -719,10 +719,10 @@ namespace TES3 {
 		bool respawnContainers = false;
 		while (gvarGameHour->value >= 24.0f) {
 			// Drop these values to integers to be later set back to the global values.
-			int day = gvarDay->value;
-			int daysPassed = gvarDaysPassed->value;
-			int month = gvarMonth->value;
-			int year = gvarYear->value;
+			int day = static_cast<int>(gvarDay->value);
+			int daysPassed = static_cast<int>(gvarDaysPassed->value);
+			int month = static_cast<int>(gvarMonth->value);
+			int year = static_cast<int>(gvarYear->value);
 
 			// Increment day count, reduce game hour by 24.
 			day++;
@@ -747,13 +747,13 @@ namespace TES3 {
 			// Do we need to respawn containers?
 			// mcp::ContainerRespawnTimescale modifies respawn cycle to use days.
 			if (advanceRespawn || mwse::mcp::getFeatureEnabled(mwse::mcp::feature::ContainerRespawnTimescale)) {
-				int monthsToRespawn = --gvarMonthsToRespawn->value;
+				int monthsToRespawn = static_cast<int>(--gvarMonthsToRespawn->value);
 				if (monthsToRespawn <= 0) {
 					respawnContainers = true;
-					gvarMonthsToRespawn->value = ndd->GMSTs[TES3::GMST::iMonthsToRespawn]->value.asLong;
+					gvarMonthsToRespawn->value = static_cast<float>(ndd->GMSTs[TES3::GMST::iMonthsToRespawn]->value.asLong);
 				}
 				else {
-					gvarMonthsToRespawn->value = monthsToRespawn;
+					gvarMonthsToRespawn->value = static_cast<float>(monthsToRespawn);
 				}
 			}
 
@@ -764,10 +764,10 @@ namespace TES3 {
 			}
 
 			// Update global variables.
-			gvarDay->value = day;
-			gvarDaysPassed->value = daysPassed;
-			gvarMonth->value = month;
-			gvarYear->value = year;
+			gvarDay->value = static_cast<float>(day);
+			gvarDaysPassed->value = static_cast<float>(daysPassed);
+			gvarMonth->value = static_cast<float>(month);
+			gvarYear->value = static_cast<float>(year);
 		}
 
 		if (respawnContainers) {
